Added width and height of image to be stored with each puzzle metadata. Took a frist crack at puzzle piece design. Got the outline to work, but using multiple 'starting points' for the path. My hope was as long as the lines touched that fill work work properly, but sadly doesn't.
I'll need to draw the piece continually either clockwise or cc. This is tricky because all puzzle pieces share 2-4 edges with other pieces. So to use the same cubic bezier curves, it has to draw in the same direction. However, top/bottom and left/right draw in opposite directions. So I need to figure out how to reverse the cubic bezier curves. Or if I alternate drawing pieces clockwise and cc, the shared edges could be drawn the same, but not sure I like that. To be continued...